optimized direct_modeling function

+++ project/direct_modeling_function/direct_modeling.py
@@ -18,38 +18,52 @@ directory = 'solutions'
 # model = .DAT ТАБЛИЧКА, ФОРМАТ: 'model_{number}.dat'
 # number = ПОРЯДКОВЫЙ НОМЕР МОДЕЛЬКИ, number = {1,2,..,10000+}
 # verbose = 0 (1, ЧТОБЫ ПОКАЗАТЬ КАРТИНКИ)
-# directory = ОПИСАНО ВЫШЕ
+# x_f, x_l, z_f, z_l, nx, nz = ПАРАМЕТРЫ ПРЯМОУГОЛЬНОЙ МОДЕЛИ (СМ. НИЖЕ)
+# cable_length = ДЛИНА КОСЫ (М)
+# elecs_step = ШАГ М/У ЭЛЕКТРОДАМИ
 
 # ФОРМАТ ФАЙЛОВ РЕШЕНИЙ -- 'origin_data_{number}.dat'
 
+#############################################################################################
 
-def direct_model(model, number, verbose):
-    L = 580                        # ДЛИНА ПРОФИЛЯ
-    l = 150                        # ДЛИНА КОСЫ
-    h = 10                         # ШАГ М/У ЭЛЕКТРОДАМИ
-    N_l = int(l / h + 1)           # КОЛ-ВО ЭЛЕКТРОДОВ В КОСЕ
-    N_L = int(L/h + 1)             # КОЛ-ВО ЭЛЕКТРОДОВ НА ПРОФИЛЕ
-    x0 = 0                         # НАЧАЛО ПРОФИЛЯ ИЗМЕРЕНИЙ (НЕ ДОЛЖЕН ВЫЛЕЗАТЬ ЗА ПРЕДЕЛЫ МОДЕЛИ!)
+# ПАРАМЕТРЫ ПРЯМОУГОЛЬНОЙ МОДЕЛИ:
+X_f = 0
+X_l = 580
+Z_f = 0
+Z_l = -60.4
+# dx = (x_l - x_f)/(nx-1)
+Nx = 233
+Nz = 41
+
+# ПАРАМЕТРЫ КОСЫ
+elecs_step = 15
+cable_length = 150
+
+
+def direct_model(model, x_f, x_l, z_f, z_l, nx, nz, elecs_step, cable_length, number, verbose):
+
+    n_elecs_cable = int(cable_length / elecs_step + 1)
+    n_elecs_profile = int((x_l - x_f) / elecs_step + 1)
 
     # СЧИТКА ДАННЫХ МОДЕЛИ ИЗ ТАБЛИЦЫ .CSV
     data = pd.read_csv(model, usecols=['#X', 'Z', 'log_Rho', 'Rho'])
-    X = np.linspace(0, 580, 233)
-    Z = np.linspace(-60.4, 0, 41)
+    X = np.linspace(x_f, x_l, nx)
+    Z = np.linspace(z_f, z_l, nz)
     R = np.array(data['Rho'])
 
     # КОЛДУНСТВО ДЛЯ ПЕРЕПИСЫВАНИЯ ВЕКТОРА СОПРОТИВЛЕНИЙ
-    Rho = np.zeros(shape=(233, 41))
+    Rho = np.zeros(shape=(nx, nz))
     for i in range(233):
-        Rho[i] = R[41*i:41*(i+1)][::-1]
+        Rho[i] = R[nz*i:nz*(i+1)]
     Rho = np.transpose(Rho)
 
     for i in range(41):
-        R[233*i:233*(i+1)] = Rho[i]
+        R[nx*i:nx*(i+1)] = Rho[i]
 
     # ГОТОВО
     Rho = R
 
-    # ГЕНЕРИМ СЕТОЧКУ ДЛЯ PYGIMLI (ПОДДЕРЖИВАЕТСЯ NUMPY)
+    # ГЕНЕРИМ СЕТОЧКУ ДЛЯ PYGIMLI
     grid = mt.createGrid(x=X, y=Z)
 
     # КОНВЕРТАЦИЯ ЗНАЧЕНИЙ В УЗЛАХ В ЗНАЧЕНИЯ ПО ПРЯМОУГОЛЬНИЧКАМ СЕТКИ
@@ -60,7 +74,7 @@ def direct_model(model, number, verbose):
     pg.show(grid, new_Rho)
 
     # СХЕМА ДИПОЛЬ-ДИПОЛЬ, расст. м/у эл. = h (м)
-    scheme = ert.createERTData(elecs=np.linspace(start=x0, stop=x0+L, num=N_L), schemeName='dd')
+    scheme = ert.createERTData(elecs=np.linspace(start=x_f, stop=x_l, num=n_elecs_profile), schemeName='dd')
 
     # МОДЕЛИРУЕМ ДАННЫЕ ПРЯМОЙ ЗАДАЧИ, С ШУМОМ
     data = ert.simulate(grid, scheme=scheme, res=new_Rho, noiseLevel=1, noiseAbs=1e-6, seed=1337)
@@ -83,7 +97,7 @@ def direct_model(model, number, verbose):
     # ПРОЧЕСЫВАЕМ ТАБЛИЧКУ .DAT, ЧТОБЫ ОБРЕЗАТЬ ТРЕУГОЛЬНИК ДО ТРАПЕЦИИ.
     # РАССТОЯНИЕ МЕЖДУ 'a' И 'n' ЭЛЕКТРОДАМИ НЕ ДОЛЖНО ПРЕВЫШАТЬ ДЛИНУ КОСЫ (N)
     for k in range(len(data['valid'])):
-        if data['n'][k] - data['a'][k] > N_l - 1:
+        if data['n'][k] - data['a'][k] > n_elecs_cable - 1:
             data['valid'][k] = 0
 
     data.save(outname)
@@ -92,7 +106,7 @@ def direct_model(model, number, verbose):
     # ПОКАЗЫВАЕМ РЕЗУЛЬТАТ ПРЯМОЙ ЗАДАЧИ
     ert.show(data)
 
-    if verbose == 1:
+    if verbose:
         plt.show()
 
     return()
@@ -100,12 +114,14 @@ def direct_model(model, number, verbose):
 
 # СОЗДАНИЕ ПАПКИ, КУДА СКЛАДЫВАТЬ ТАБЛИЧКИ С ПРЯМЫМИ ЗАДАЧАМИ.
 # ЕСЛИ ОНА УЖЕ СУЩЕСТВУЕТ, УДАЛЯЕМ И СОЗДАЕМ ЗАНОВО:
-shutil.rmtree(directory, ignore_errors=True)
+if os.path.exists(directory):
+    shutil.rmtree(directory, ignore_errors=True)
 os.mkdir(directory)
 
 # ИНИЦИИРУЕМ ФУНКЦИЮ ДЛЯ МОДЕЛЕЙ № 9,15,16:
 for i in [0]:
     os.chdir('models')
     filename = 'model_' + str(i) + '.csv'
-    direct_model(model=filename, verbose=1, number=i)
+    direct_model(model=filename, cable_length=cable_length, elecs_step=elecs_step, x_f=X_f, x_l=X_l, z_f=Z_f, z_l=Z_l,
+                 nx=Nx, nz=Nz, verbose=True, number=i)
     os.chdir('..')
